#12278 move IPAddressField to field_serializers
